# docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: Kanbien Platform API
  version: 0.2.0
  description: >
    OpenAPI specification for the platform API. This documents the platform
    health route, root-user authentication, root-role authorization
    administration, and the protected rootUsers feature.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/root-users:
    post:
      tags:
        - Root Users
      summary: Create a root user
      operationId: createRootUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRootUserBody"
      responses:
        "201":
          description: Root user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags:
        - Root Users
      summary: List root users or look up by exact email
      operationId: queryRootUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/RootUsersOrderBy"
        - $ref: "#/components/parameters/OrderDirection"
        - $ref: "#/components/parameters/Email"
        - $ref: "#/components/parameters/EmailPrefix"
        - $ref: "#/components/parameters/FirstNamePrefix"
        - $ref: "#/components/parameters/LastNamePrefix"
        - $ref: "#/components/parameters/CreatedAtFrom"
        - $ref: "#/components/parameters/CreatedAtTo"
        - $ref: "#/components/parameters/UpdatedAtFrom"
        - $ref: "#/components/parameters/UpdatedAtTo"
        - $ref: "#/components/parameters/DeletedAtFrom"
        - $ref: "#/components/parameters/DeletedAtTo"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Exact-match root user or paginated list result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RootUser"
                  - $ref: "#/components/schemas/PaginatedRootUsers"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-users/active:
    get:
      tags:
        - Root Users
      summary: List active root users
      operationId: listActiveRootUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/ActiveRootUsersOrderBy"
        - $ref: "#/components/parameters/OrderDirection"
        - $ref: "#/components/parameters/EmailPrefix"
        - $ref: "#/components/parameters/FirstNamePrefix"
        - $ref: "#/components/parameters/LastNamePrefix"
        - $ref: "#/components/parameters/CreatedAtFrom"
        - $ref: "#/components/parameters/CreatedAtTo"
        - $ref: "#/components/parameters/UpdatedAtFrom"
        - $ref: "#/components/parameters/UpdatedAtTo"
      responses:
        "200":
          description: Paginated active root users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRootUsers"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-users/deleted:
    get:
      tags:
        - Root Users
      summary: List deleted root users
      operationId: listDeletedRootUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/RootUsersOrderBy"
        - $ref: "#/components/parameters/OrderDirection"
        - $ref: "#/components/parameters/EmailPrefix"
        - $ref: "#/components/parameters/FirstNamePrefix"
        - $ref: "#/components/parameters/LastNamePrefix"
        - $ref: "#/components/parameters/CreatedAtFrom"
        - $ref: "#/components/parameters/CreatedAtTo"
        - $ref: "#/components/parameters/UpdatedAtFrom"
        - $ref: "#/components/parameters/UpdatedAtTo"
        - $ref: "#/components/parameters/DeletedAtFrom"
        - $ref: "#/components/parameters/DeletedAtTo"
        - $ref: "#/components/parameters/ExcludeAnonymized"
      responses:
        "200":
          description: Paginated deleted root users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRootUsers"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-users/{rootUserId}:
    get:
      tags:
        - Root Users
      summary: Get a visible root user by ID
      operationId: getRootUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RootUserId"
      responses:
        "200":
          description: Root user found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags:
        - Root Users
      summary: Update a visible root user
      operationId: updateRootUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RootUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRootUserBody"
      responses:
        "200":
          description: Root user updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Root Users
      summary: Soft delete a root user
      operationId: deleteRootUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RootUserId"
      responses:
        "200":
          description: Root user soft deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-users/{rootUserId}/remove:
    post:
      tags:
        - Root Users
      summary: Anonymize and remove a root user
      operationId: removeRootUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RootUserId"
      responses:
        "200":
          description: Root user anonymized and removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/root-users/{rootUserId}/reactivate:
    post:
      tags:
        - Root Users
      summary: Reactivate a soft-deleted root user
      operationId: reactivateRootUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RootUserId"
      responses:
        "200":
          description: Root user reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
    RootUserId:
      name: rootUserId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    OrderDirection:
      name: orderDirection
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    RootUsersOrderBy:
      name: orderBy
      in: query
      required: false
      schema:
        type: string
        enum: [email, firstName, lastName, status, createdAt, updatedAt, deletedAt]
        default: updatedAt
    ActiveRootUsersOrderBy:
      name: orderBy
      in: query
      required: false
      schema:
        type: string
        enum: [email, firstName, lastName, createdAt, updatedAt]
        default: updatedAt
    Email:
      name: email
      in: query
      required: false
      schema:
        type: string
        format: email
    EmailPrefix:
      name: emailPrefix
      in: query
      required: false
      schema:
        type: string
        minLength: 3
    FirstNamePrefix:
      name: firstNamePrefix
      in: query
      required: false
      schema:
        type: string
        minLength: 3
    LastNamePrefix:
      name: lastNamePrefix
      in: query
      required: false
      schema:
        type: string
        minLength: 3
    CreatedAtFrom:
      name: createdAtFrom
      in: query
      required: false
      schema:
        type: string
        format: date-time
    CreatedAtTo:
      name: createdAtTo
      in: query
      required: false
      schema:
        type: string
        format: date-time
    UpdatedAtFrom:
      name: updatedAtFrom
      in: query
      required: false
      schema:
        type: string
        format: date-time
    UpdatedAtTo:
      name: updatedAtTo
      in: query
      required: false
      schema:
        type: string
        format: date-time
    DeletedAtFrom:
      name: deletedAtFrom
      in: query
      required: false
      schema:
        type: string
        format: date-time
    DeletedAtTo:
      name: deletedAtTo
      in: query
      required: false
      schema:
        type: string
        format: date-time
    Status:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RootUserStatus"
    ExcludeAnonymized:
      name: excludeAnonymized
      in: query
      required: false
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or was not supplied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict with existing or current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Request was throttled or temporarily locked due to too many attempts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RootUserStatus:
      type: string
      enum: [active, inactive]
...
    CountValue:
      oneOf:
        - type: integer
        - type: string
          enum: ["10000+"]
    RootUser:
      type: object
      properties:
        rootUserId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        anonymized:
          type: boolean
        status:
          $ref: "#/components/schemas/RootUserStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, "null"]
          format: date-time
      required:
        - rootUserId
        - email
        - anonymized
        - status
        - createdAt
        - updatedAt
        - deletedAt
    PaginatedRootUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RootUser"
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        totalSearchableRecords:
          $ref: "#/components/schemas/CountValue"
        totalMatchingRecords:
          $ref: "#/components/schemas/CountValue"
      required:
        - items
        - page
        - pageSize
        - totalPages
        - totalSearchableRecords
        - totalMatchingRecords
...
    CreateRootUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
      required: [email]
    UpdateRootUserBody:
      type: object
      minProperties: 1
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/RootUserStatus"
...
    ErrorDetails:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          $ref: "#/components/schemas/ErrorDetails"
      required: [code, message]